Updated Background.md and Sprites.md
Through experimentation in gb-studio I noticed my sprites had drastically changed their colors because I used #306850 in my spritesheets. This is a color that only backgrounds can have, and this is not pointed out in the docs as a limitation of spritesheets. I added and changed a bit of the text on background.md and sprites.md that identifies this limitation of spritesheets and backgrounds so developers are aware they can't use #306850 on sprites or #65FF00 on backgrounds.