Updated backgrounds.md and sprites.md
I was lead to believe that #306850 was a color usable on spritesheets. After losing my sprite colors in-game too many times, I realized that I was using #306850 on sprites when it is only listed in the backgrounds palette. I felt this was not pointed out well in the docs, so I added some text in these two doc pages to make it obvious that #306850 can not be used on sprites and #65FF00 can not be used on backgrounds.