+Volume Flooring info, Correct effect usage, formatting
- Renamed "Important Limitations" to "GBT Player's Channel Limitations"

- Reordered Effects to be in order of "Note Effects" followed by "Command Effects"

Going off of how rulz explained these effects, I added this distinction to describe the sole rule that Channel 3 has about its effect limits. Previously we thought that Channel 3 was incapable of any effects other than 0xy, E8x, ECx and Cxx. The previous documentation even forgot to include that Channel 3 can use effect Cxx.

- I re-wrote the Note Range footnote on the basis of One-Indexing and Zero-Indexing

This is so the information translates across to more trackers, and using the currently popular OpenMPT and MilkyTracker as examples to follow.

- Volume "flooring" to multiples of 4 and Channel 3's volume "rounding" were finally covered (thanks rulz!)

I also made it clear that non-distinct volume numbers for the channels can still be entered without drastic problems. I also made sure to include base 10 and hexadecimal examples to accomodate for all users, and gave examples in base 10.

- Added the "Volume Persistence" heading to emninem's examples

@emninem did a great job on this example. I did expand on a few of the potentially-ambiguous uses of "this" and "that", and avoided using contradictory language like "meanwhile".

- Renamed "Instruments available" to "Instruments"

I also removed the segment I wrote about how I'm denoting hexadecimal numbers. I instead opted for `Base-10-First. (0hx following)` by adding "hx" to the end of the hexadecimal instrument numbers that I'd previously added. The description for instrument limits was reworked. Example:

``The wave channel gets 8 instruments (from 8 to 15):`` is now ``Channel 3, the wave channel, has 8 instrument options:``. The numbers of those instruments are already listed below as-needed.

Any cases of describing the noise instruments as having "The same waveform but faster" I mostly changed to "etc." to make the list easier to parse. This is still not a great description system and I'm convinced that someone other than me needs to make this part easier to skim because I haven't found the best solution yet.

- Updating the old monospace text formatting of AntonioND's Effects table with Github Markdown formatting (finally!)

I also split up the effects into "Note-effects" and "Command-effects" going off of how rulz described it. This is a lot easier to compartmentalize, especially as Channel 3's effect restrictions only apply to Command effects. This note is included below the Command-effect table.

I also tried to cover what each effect does in clearer words. Note-Cut's old description of `0 < x < speed-1` only made sense to me just now despite reading the docs for almost a year, so now Note Cut's requirements are written in words rather than a math equation.

- Updating some FAQs with new information

We now have sound effects! I felt this was important to emphasize in the FAQs so as to not completely remove those contents. This question may be redundant as we can just advance to the next page to show that sound-effects can be done.

The question on .midi to .mod tried to explain the process in the question's answer, and it may be best to simply link a page that explains how to do it for OpenMPT. That process should be documentation for OpenMPT, not GBT Player/GB Studio, but it's good that we still acknowledge this FAQ in this documentation.

- Removed some semantic words and phrasings

I followed these two guides: http://www.plainenglish.co.uk/files/howto.pdf and https://docs.godotengine.org/en/stable/community/contributing/docs_writing_guidelines.html as part of my reasoning. I did not follow these guides strictly, but felt it was necessary to better-convey the information that's in these docs. This included removing words like "please", "basically", and "so". Some sentences were reworked to allow the document's information to appear sooner in various sentences. For example, one of the earlier sentences used to be `There may be only 4 channels in your .MOD file.` whereas now it is `.MOD files need to use 4 channels.`