Corrected Vol. Examples to use Hex, Cleaned-Up FaQ
I edited some of the FaQ to all be in question format, and moved around some of the information in questions. I still don't think it's perfectly organized, I definitely ruined some of the original questions' intent, but hopefully only when the answer to the question that was being asked had since changed in the past few months (like .midi to .mod). I removed my part about ".wav to .midi" because nobody should ever really have to use a .wav to .midi "converter".