music doc opening is more like other doc openings
I noticed that the music doc is an untamed beast of writing conventions compared to the other docs, most likely because Chris's writing voice is distinct compared to the open-sourced result of the music docs, which is fine and expected. I still wanted to try to see what people thought if the music docs opened with a reminder of how to add music to your game, since this reminder is in every other page of the docs.