++"A: F: T:" limits in Scenes & Actors, ++Actor properties
Sorry for gutting your original descriptions about Movement Types Chris. Also, I'm not sure if this list is the best way to do Properties but Actors have the most complicated properties and I felt that it deserved them. We could always replace this big list with a good visual diagram instead.