Remove FaQ and Tips, add Converting
While the informatiMy goal is to make the whole GBT docs shorter and more useful, since it is no longer the only option for making music. While the previous FaQ and Tips sections were useful, I felt their information was either outdated, or they covered topics that would better fit in a unique article (motivation, starting small) or a specific tracker's documentation (quick envelopes). What I added in its place to keep most of its information was a section on converting other file formats to .mod, because it's still a common way that people get their music to the gbt-compatible .mod format.